[no ci] chore(devenv): get rid of it
This worked **perfectly fine** 12 hours ago. I tried to use it after a while, and I was met with the image not building. I troubleshooted it for a while and found a fix that worked. Went to get some ice cream because New Year's Eve. Came back. Tried building the image again. It doesn't work anymore. What the fuck. What did I even change